+++ core/capabilities/vault/utils/json_test.go
@@ -1,4 +1,4 @@
-package vault
+package utils
 
 import (
 	"testing"
