+++ system-tests/tests/smoke/cre/cre_suite_test.go
@@ -92,7 +92,7 @@ To execute tests with v2 contracts start the local CRE first:
 */
 func Test_CRE_V2_Suite(t *testing.T) {
 	topology := os.Getenv("TOPOLOGY_NAME")
-
+/*
 	t.Run("[v2] Proof Of Reserve - "+topology, func(t *testing.T) {
 		// TODO: Review why this test cannot run with two chains? (CRE-983)
 		// How to configure evm for both chains and capabilities DON (DON<>DON topology)?
@@ -104,13 +104,13 @@ func Test_CRE_V2_Suite(t *testing.T) {
 		wfConfig.FeedIDs = []string{wfConfig.FeedIDs[0]}
 		ExecutePoRTest(t, testEnv, priceProvider, wfConfig, false)
 	})
-
+*/
 	t.Run("[v2] Vault DON - "+topology, func(t *testing.T) {
 		testEnv := t_helpers.SetupTestEnvironmentWithConfig(t, t_helpers.GetDefaultTestConfig(t), v2RegistriesFlags...)
 
 		ExecuteVaultTest(t, testEnv)
 	})
-
+/*
 	t.Run("[v2] Cron Beholder - "+topology, func(t *testing.T) {
 		testEnv := t_helpers.SetupTestEnvironmentWithConfig(t, t_helpers.GetDefaultTestConfig(t), v2RegistriesFlags...)
 
@@ -134,8 +134,10 @@ func Test_CRE_V2_Suite(t *testing.T) {
 
 		ExecuteConsensusTest(t, testEnv)
 	})
-}
 
+ */
+}
+/*
 func Test_CRE_V2_EVM_Suite(t *testing.T) {
 	topology := os.Getenv("TOPOLOGY_NAME")
 	testEnv := t_helpers.SetupTestEnvironmentWithConfig(t, t_helpers.GetDefaultTestConfig(t), v2RegistriesFlags...)
@@ -152,3 +154,4 @@ func Test_CRE_V2_EVM_Suite(t *testing.T) {
 		ExecuteEVMReadTest(t, testEnv)
 	})
 }
+*/
